+++ src/main/java/redis/clients/jedis/ShardedJedis.java
@@ -214,27 +214,27 @@ public class ShardedJedis extends BinaryShardedJedis implements JedisCommands {
         return j.llen(key);
     }
 
-    public List<String> lrange(String key, int start, int end) {
+    public List<String> lrange(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.lrange(key, start, end);
     }
 
-    public String ltrim(String key, int start, int end) {
+    public String ltrim(String key, long start, long end) {
         Jedis j = getShard(key);
         return j.ltrim(key, start, end);
     }
 
-    public String lindex(String key, int index) {
+    public String lindex(String key, long index) {
         Jedis j = getShard(key);
         return j.lindex(key, index);
     }
 
-    public String lset(String key, int index, String value) {
+    public String lset(String key, long index, String value) {
         Jedis j = getShard(key);
         return j.lset(key, index, value);
     }
 
-    public Long lrem(String key, int count, String value) {
+    public Long lrem(String key, long count, String value) {
         Jedis j = getShard(key);
         return j.lrem(key, count, value);
     }
